Move ThemedRadioButton to GeneralForm folder
