+++ src/app/settings/views/Configuration/GeneralForm/ThemedRadioButton/_index.scss
